Fix:prettier apply and name convention changing
